common: add device dax support

+++ src/tools/pmempool/rm.c
@@ -45,7 +45,7 @@
 #include "common.h"
 #include "output.h"
 #include "rm.h"
-
+#include "device_dax.h"
 #include "set.h"
 
 #ifdef USE_RPMEM
@@ -134,8 +134,14 @@ rm_file(const char *file)
 
 	const char *pre_msg = write_protected ? "write-protected " : "";
 	if (ask_Yn(cask, "remove %sfile '%s' ?", pre_msg, file) == 'y') {
-		if (unlink(file))
-			err(1, "cannot remove file '%s'", file);
+		int dax = device_dax_is_dax(file);
+		if (dax) {
+			if (device_dax_zero(file) != 0)
+				err(1, "cannot zero device dax '%s'", file);
+		} else {
+			if (unlink(file))
+				err(1, "cannot remove file '%s'", file);
+		}
 		outv(1, "removed '%s'\n", file);
 	}
 }
@@ -308,7 +314,7 @@ pmempool_rm_func(char *appname, int argc, char *argv[])
 			err(1, "cannot remove '%s'", file);
 		}
 
-		int is_poolset = util_is_poolset_file(file);
+		int is_poolset = util_is_poolset_file(file) == 1;
 
 		if (is_poolset)
 			outv(2, "poolset file: %s\n", file);
